Cast optimization improvement: also handle local.tee
Converts the following:

 (some.operation
   (ref.cast .. (local.tee $ref ..))
   (local.get $ref)
 )

into:

 (some.operation
   (local.tee $temp
     (ref.cast .. (local.tee $ref ..))
   )
   (local.get $temp)
 )

This removes close to 10% of the casts in a test program (from 38113
casts down to 36119).

I also tried to improve the optimization that moves more refined casts
earlier, but this does not seem very effective (it only eliminated 9
further casts), so I'm not sure it is worth it.