DebugLocationPropagation: pass debuglocation from parent node to chil… (#6500)
This PR creates a pass to propagate debug location from parent node to child nodes which has no debug location with pre-order traversal. This is useful for compilers that use Binaryen API to generate WebAssembly modules. 

It behaves like `wasm-opt` read text format file: children are tagged with the debug info of the parent, if they have no annotation of their own.

For compilers that use Binaryen API to generate WebAssembly modules, it is a bit redundant to add debugInfo for each expression, Especially when the compiler wrap expressions. 

With this pass, compilers just need to add debugInfo for the parent node, which is more convenient.

For example:
```
(drop
  (call $voidFunc)
)
```

Without this pass, if the compiler only adds debugInfo for the wrapped expression `drop`, the `call` expression has no corresponding source code mapping in DevTools debugging, which is obviously not user-friendly.