Source maps: allow to specify that an expression has no debug info in text parsers
The comment `;;@` with nothing else can be used to specify that the
following expression does not have any debug info associated to it. This
can be used to stop the automatic propagation of debug info in the text
parsers.

The text printer has also been updated to output this comment when needed.