[Parser] Do not eagerly lex parens (#6540)
The lexer currently lexes tokens eagerly and stores them in a `Token` variant
ahead of when they are actually requested by the parser. It is wasteful,
however, to classify tokens before they are requested by the parser because it
is likely that the next token will be precisely the kind the parser requests.
The work of checking and rejecting other possible classifications ahead of time
is not useful.

To make incremental progress toward removing `Token` completely, lex parentheses
on demand instead of eagerly.