Emit stringview conversions in binary writer
So that we continue to produce valid stringref code. Parse the conversions as
nops in the binary parser to preserve our ability to round-trip. Emitting the
conversions requires using extra scratch locals, and between that and parsing
the nop, there is quite a bit of code bloat when round-tripping, but stringref
code should never be emitted in production use cases, so that's ok.