LocalCSE: Ignore traps of code in between (#6588)
Given:

(ORIGINAL)
(in between)
(COPY)

We want to change that to

(local.tee $temp (ORIGINAL))
(in between)
(local.get $temp)

It is fine if "in between" traps: then we never reach the new local.get. This is a safer
situation than most optimizations because we are not reordering anything, only
replacing known-equivalent code.