Fix wasm-split bug in absence of active element segments (#6651)
The module splitting code incorrectly assumed that there would be at least one
active element segment and failed to initialize the table slot manager with a
function table if that was not the case. Fix the bug by setting the table even
when there are no active segments and add a test.

Fixes #6572 and #6637.