Add local.set/tee local type annotations to BINARYEN_PRINT_FULL  (#6657)
With this we now print e.g.

(local.set $temp (; local type: i32 ;)
  ...

This can be nice in large functions to avoid needing to scroll up to
see the local type, e.g. when debugging why unsubtyping doesn't
work somewhere.

Also avoid [ ] in this mode, in favor of the standard (; ;), and put those
at the end rather than at the start.