[threads] Shared basic heap types
Implement binary and text parsing and printing of shared basic heap types and
incorporate them into the type hierarchy.

To avoid the massive amount of code duplication that would be necessary if we
were to add separate enum variants for each of the shared basic heap types, use
bit 0 to indicate whether the type is shared and replace `getBasic()` with
`getUnsharedBasic()`, which clears that bit. Update all the use sites to record
whether the original type was shared and produce shared or unshared output
without code duplication.