GlobalStructInference: Un-nest struct.news in globals when that is helpful (#6688)
If we have

(global $g (struct.new $S
  (i32.const 1)
  (struct.new $T ..)
  (ref.func $f)
))

then before this PR if we wanted to read the middle field we'd stop, as it is non-constant.
However, we can un-nest it, making it constant:

(global $g.unnested (struct.new $T ..))
(global $g (struct.new $S
  (i32.const 1)
  (global.get $g.unnested)
  (ref.func $f)
))

Now the field is a global.get of an immutable global, which is constant. Using this
technique we can handle anything in a struct field, constant or not. The cost of adding
a global is likely offset by the benefit of being able to refer to it directly, as that opens
up more opportunities later.

Concretely, this replaces the constant values we look for in GSI with a variant over
constants or expressions (we do still want to group constants, as multiple globals
with the same constant field can be treated as a whole). And we note cases where we
need to un-nest, and handle those at the end.